backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sentence-transformers==2.2.2
faiss-cpu>=1.8.0
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.3.0
langchain==0.0.350
langchain-community==0.0.10
numpy>=1.26.0
python-dotenv==1.0.0

